+++ src/PDCTaskRunner.ts
@@ -4,15 +4,12 @@ import { ConfigurationResolver } from "./ConfigurationResolver";
 import { TaskRunner } from "./TaskRunner";
 import { exec } from "./exec";
 import { quote } from "./quote";
-import { wait } from "./wait";
 
 /**
  * PDCTaskRunnerOptions contains extra properties asigned to the `pdc` task in
  * `tasks.json`.
  */
-export interface PDCTaskRunnerOptions {
-  timeout?: number;
-}
+export interface PDCTaskRunnerOptions {}
 
 /**
  * PDCTaskRunner is responsible for executing the PlaydateSDK's `pdc` binary
@@ -27,18 +24,11 @@ export class PDCTaskRunner implements TaskRunner {
   ) {}
 
   async run(): Promise<void> {
-    const { timeout } = this.options;
     const { sdkPath, sourcePath, productPath } = await this.config.resolve();
 
     const cmd = path.join(sdkPath, "bin", "pdc");
     const args = [quote(sourcePath), quote(productPath)];
     const command = `${cmd} ${args.join(" ")}`;
     await exec(command);
-
-    if (!timeout) {
-      return;
-    }
-
-    await wait(timeout);
   }
 }